+++ packages/e2e.sandbox.officenetwork/src/generatedNoCheck2/ontology/interfaces.ts
@@ -1,0 +1,4 @@
+export { ExampleInterface } from './interfaces/ExampleInterface.js';
+export { Person } from './interfaces/Person.js';
+export { TodoLike } from './interfaces/TodoLike.js';
+export { Worker } from './interfaces/Worker.js';