+++ benchmarks/tests/primary/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/benchmarks.primary
 
+## 0.1.0-beta.101
+
+### Patch Changes
+
+- @osdk/client@2.8.0-beta.2
+
 ## 0.1.0-beta.100
 
 ### Patch Changes