+++ packages/client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @osdk/client
 
+## 2.8.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [e462938]
+  - @osdk/client.unstable@2.8.0-beta.2
+  - @osdk/api@2.8.0-beta.2
+  - @osdk/generator-converters@2.8.0-beta.2
+
 ## 2.7.0-beta.14
 
 ### Minor Changes