+++ packages/create-widget.template.react.v2/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @osdk/create-widget.template.react.v2
 
+## 3.5.0-beta.2
+
 ## 3.4.0-beta.14
 
 ## 3.4.0-beta.13