+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @osdk/generator
 
+## 2.8.0-beta.2
+
+### Minor Changes
+
+- 1706c79: Relax interface creation validation.
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.2
+- @osdk/generator-converters@2.8.0-beta.2
+
 ## 2.7.0-beta.14
 
 ### Patch Changes