+++ packages/vite-plugin-oac/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @osdk/vite-plugin-oac
 
+## 0.6.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [e462938]
+  - @osdk/client.unstable@2.8.0-beta.2
+  - @osdk/maker@0.16.0-beta.2
+  - @osdk/api@2.8.0-beta.2
+  - @osdk/generator-converters.ontologyir@2.8.0-beta.2
+
 ## 0.5.0-beta.14
 
 ### Patch Changes