+++ packages/widget.api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/widget.api
 
+## 3.5.0-beta.2
+
+### Patch Changes
+
+- @osdk/api@2.8.0-beta.2
+
 ## 3.4.0-beta.14
 
 ### Patch Changes